2/11 14:26 박명환 - 미션 부분과 OpenAI Long -> Integer 수정

+++ src/main/java/com/example/team5_be/openai/ctrl/OpenAIController.java
@@ -35,7 +35,7 @@ public class OpenAIController {
     )
     @GetMapping("/recommend/mission")
     public MissionRecommendationResponseDTO recommendMission(
-        @RequestParam Long habitId,
+        @RequestParam Integer habitId,
         @RequestParam String userId
     ) {
         return openAIService.recommendMission(habitId, userId);
